Update build to include setting the version
@DLWyatt if you don't like this I can revert this, but @JPRuskin is right that this is how I prefer to do it

+++ Build.ps1
@@ -1,12 +1,29 @@
 [CmdletBinding()]
-param($OutputDirectory = $("$PSScriptRoot\Output"), [switch]$Force)
+param(
+    # A version number to update the output with (uses gitversion by default)
+    $Version = $(gitversion -showvariable nugetversion),
 
+    # The output folder (defaults to the version number)
+    $OutputDirectory = $("$PSScriptRoot\$(($Version -split '[-+]',2)[0])"),
 
-Remove-Item $OutputDirectory -Recurse -Confirm:(!$Force)
+    # If set, removes the output folder without prompting!
+    [switch]$Force
+)
 
+$VersionPrefix, $VersionSuffix = $Version -split '[-+]', 2
 
-Copy-Item $PSScriptRoot\Module -Destination $OutputDirectory -recurse
+if (Test-Path $OutputDirectory) {
+    Remove-Item $OutputDirectory -Recurse -Confirm:(!$Force)
+}
 
-dotnet build -c Release
+Copy-Item $PSScriptRoot\Module\PowerShellLogging -Destination $OutputDirectory -recurse
+Set-Content $OutputDirectory\PowerShellLogging.psd1 (
+    (Get-Content $OutputDirectory\PowerShellLogging.psd1) -replace
+        "(ModuleVersion\s+=\s+)'.*'", "`$1'$VersionPrefix'" -replace
+        "(Prerelease\s+=\s+)'.*'", "`$1'$VersionSuffix'"
+)
+
+dotnet build -c Release -p:VersionPrefix=$VersionPrefix -p:VersionSuffix=$VersionSuffix
+
 Get-ChildItem bin/Release -recurse -filter *.dll |
-    Move-Item -Destination $OutputDirectory\PowerShellLogging\
+    Move-Item -Destination $OutputDirectory\
